Add files via upload
DESCRIPTION : 
FLAMES is a popular game that predicts the relationship between two individuals based on their names. The acronym "FLAMES" stands for Friendship, Love, Affection, Marriage, Enmity, and Siblings. The game is often played for fun and entertainment, particularly among young people.

Here's how the FLAMES game typically works:

Input Names: The game starts by taking the names of the two individuals whose relationship is to be predicted. Each individual's name is typically represented as a string of characters.

Remove Common Letters: Common letters (including duplicates) between the two names are eliminated. This step ensures that each letter is counted only once in the calculation.

Count Remaining Letters: After removing common letters, the remaining letters in both names are counted.

Calculate FLAMES: The total count of the remaining letters is then used to determine the relationship between the two individuals. The count is matched with the letters in the word "FLAMES". Each letter in "FLAMES" represents a relationship category:

F: Friendship
L: Love
A: Affection
M: Marriage
E: Enmity
S: Siblings

if the count is greater than the number of letters in Flames then it is cycled again from the first letter of FLAMES.

Result: The letter at which the count ends corresponds to the relationship category, indicating the type of relationship between the two individuals.

For example, if the count results in the letter "M", it suggests that the relationship between the individuals could lead to marriage. If the count results in the letter "F", it indicates a friendly relationship. 

Feel free to run the code and test your relationships!!!